vit-main/requirements.txt
Flask
torch
transformers
Pillow
numpy
lime
scikit-image
matplotlib
Werkzeug
